refactor(smokeynet-adapted): split tubes.py, dedup params, drop dead helpers
- tubes.py (1057 LOC) was doing two jobs: construction and rendering. Pull
  the ~500 LOC of matplotlib/cv2-based visualization out into tube_viz.py
  so construction has no heavy plotting imports.
- Drop classify_tube_gt, compute_containment, and load_gt_labels. They had
  tests but no runtime callers (last user was the deleted SmokeyNetAdapted
  scoring path).
- params.yaml: seven train_* variants collapsed onto YAML anchors. Resolved
  values per variant are byte-identical to the hand-rolled version, so no
  dvc.lock churn. Verified that DVC's own load_yaml expands <<: *anchor.
- Update scripts/render_tubes.py, notebook 02, dvc.yaml deps, and test
  imports to the new module split. Add historical notes at the top of
  the two design docs that reference the now-deleted architecture.
